changing type of a parameter (spotted during an interview)

+++ task-data/src/main/java/org/vaadin/hr/datareader/data/Person.java
@@ -52,7 +52,7 @@ public final class Person {
         return age;
     }
 
-    public void setAge(byte age) {
+    public void setAge(int age) {
         this.age = age;
     }
